+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.3.1] / 2024-04-02
+### Fixed
+- Fix `Installation` Application issue.
+### Updated
+- Update `ricaun.Revit.Installation`
+
 ## [1.3.0] / 2024-03-28
 ### Features
 - Readme information to wiki.